Moved weapons and game_key outside of shoot() scope and added typing hints.
The weapon data is immutable between calls to shoot/{weapon} endpoint so this should reduce response delays.

FastAPI makes extensive use of Pydantic so adding typing is reasonable.

Moved opp_weapon outside of shoot(), reduces knowledge of internal data from the endpoint routine to facilitate moving endpoints outside main when reasonable. This was not awaitable as it is considered atomic to any endpoint calling it.

Started adding docstrings.